refactor bootstrap token utils
