+++ cmd/kubeadm/app/apis/kubeadm/bootstraptokenhelpers_test.go
@@ -454,55 +454,3 @@ func jsonMarshal(bt *BootstrapToken) string {
 	b, _ := json.Marshal(*bt)
 	return string(b)
 }
-
-func TestGetSecretString(t *testing.T) {
-	var tests = []struct {
-		name        string
-		secret      *v1.Secret
-		key         string
-		expectedVal string
-	}{
-		{
-			name: "existing key",
-			secret: &v1.Secret{
-				ObjectMeta: metav1.ObjectMeta{Name: "foo"},
-				Data: map[string][]byte{
-					"foo": []byte("bar"),
-				},
-			},
-			key:         "foo",
-			expectedVal: "bar",
-		},
-		{
-			name: "non-existing key",
-			secret: &v1.Secret{
-				ObjectMeta: metav1.ObjectMeta{Name: "foo"},
-				Data: map[string][]byte{
-					"foo": []byte("bar"),
-				},
-			},
-			key:         "baz",
-			expectedVal: "",
-		},
-		{
-			name: "no data",
-			secret: &v1.Secret{
-				ObjectMeta: metav1.ObjectMeta{Name: "foo"},
-			},
-			key:         "foo",
-			expectedVal: "",
-		},
-	}
-	for _, rt := range tests {
-		t.Run(rt.name, func(t *testing.T) {
-			actual := getSecretString(rt.secret, rt.key)
-			if actual != rt.expectedVal {
-				t.Errorf(
-					"failed getSecretString:\n\texpected: %s\n\t  actual: %s",
-					rt.expectedVal,
-					actual,
-				)
-			}
-		})
-	}
-}
